Code factorisation for rendering system.
